Change List from common to components and rename it TopicList
